+++ lib/js_of_ocaml-runtime/js_of_ocaml_runtime_stubs.c
@@ -1,0 +1,110 @@
+#include <stdlib.h>
+#include <stdio.h>
+void caml_exn_with_js_backtrace () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_exn_with_js_backtrace!\n");
+  exit(1);
+}
+void caml_js_call () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_call!\n");
+  exit(1);
+}
+void caml_js_delete () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_delete!\n");
+  exit(1);
+}
+void caml_js_equals () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_equals!\n");
+  exit(1);
+}
+void caml_js_error_of_exception () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_error_of_exception!\n");
+  exit(1);
+}
+void caml_js_eval_string () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_eval_string!\n");
+  exit(1);
+}
+void caml_js_expr () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_expr!\n");
+  exit(1);
+}
+void caml_js_fun_call () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_fun_call!\n");
+  exit(1);
+}
+void caml_js_get () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_get!\n");
+  exit(1);
+}
+void caml_js_meth_call () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_meth_call!\n");
+  exit(1);
+}
+void caml_js_new () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_new!\n");
+  exit(1);
+}
+void caml_js_object () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_object!\n");
+  exit(1);
+}
+void caml_js_pure_expr () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_pure_expr!\n");
+  exit(1);
+}
+void caml_js_set () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_set!\n");
+  exit(1);
+}
+void caml_js_var () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_var!\n");
+  exit(1);
+}
+void caml_js_wrap_callback () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback!\n");
+  exit(1);
+}
+void caml_js_wrap_callback_arguments () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback_arguments!\n");
+  exit(1);
+}
+void caml_js_wrap_callback_strict () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback_strict!\n");
+  exit(1);
+}
+void caml_js_wrap_meth_callback () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback!\n");
+  exit(1);
+}
+void caml_js_wrap_meth_callback_arguments () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_arguments!\n");
+  exit(1);
+}
+void caml_js_wrap_meth_callback_strict () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_strict!\n");
+  exit(1);
+}
+void caml_js_wrap_meth_callback_unsafe () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_unsafe!\n");
+  exit(1);
+}
+void caml_jsstring_of_string () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_jsstring_of_string!\n");
+  exit(1);
+}
+void caml_ojs_new_arr () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_ojs_new_arr!\n");
+  exit(1);
+}
+void caml_pure_js_expr () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_pure_js_expr!\n");
+  exit(1);
+}
+void caml_string_of_jsstring () {
+  fprintf(stderr, "Unimplemented Javascript primitive caml_string_of_jsstring!\n");
+  exit(1);
+}
+void debugger () {
+  fprintf(stderr, "Unimplemented Javascript primitive debugger!\n");
+  exit(1);
+}
