+++ lib/js_of_ocaml/js_of_ocaml_stubs.c
@@ -32,36 +32,8 @@ void caml_create_file () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_create_file!\n");
   exit(1);
 }
-void caml_exn_with_js_backtrace () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_exn_with_js_backtrace!\n");
-  exit(1);
-}
 void caml_int64_create_lo_mi_hi () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_int64_create_lo_mi_hi!\n");
-  exit(1);
-}
-void caml_js_call () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_call!\n");
-  exit(1);
-}
-void caml_js_delete () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_delete!\n");
-  exit(1);
-}
-void caml_js_equals () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_equals!\n");
-  exit(1);
-}
-void caml_js_error_of_exception () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_error_of_exception!\n");
-  exit(1);
-}
-void caml_js_eval_string () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_eval_string!\n");
-  exit(1);
-}
-void caml_js_expr () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_expr!\n");
   exit(1);
 }
 void caml_js_from_array () {
@@ -80,14 +52,6 @@ void caml_js_from_string () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_from_string!\n");
   exit(1);
 }
-void caml_js_fun_call () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_fun_call!\n");
-  exit(1);
-}
-void caml_js_get () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_get!\n");
-  exit(1);
-}
 void caml_js_get_console () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_get_console!\n");
   exit(1);
@@ -104,28 +68,8 @@ void caml_js_instanceof () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_instanceof!\n");
   exit(1);
 }
-void caml_js_meth_call () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_meth_call!\n");
-  exit(1);
-}
-void caml_js_new () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_new!\n");
-  exit(1);
-}
-void caml_js_object () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_object!\n");
-  exit(1);
-}
 void caml_js_on_ie () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_on_ie!\n");
-  exit(1);
-}
-void caml_js_pure_expr () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_pure_expr!\n");
-  exit(1);
-}
-void caml_js_set () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_set!\n");
   exit(1);
 }
 void caml_js_to_array () {
@@ -152,44 +96,8 @@ void caml_js_typeof () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_typeof!\n");
   exit(1);
 }
-void caml_js_var () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_var!\n");
-  exit(1);
-}
-void caml_js_wrap_callback () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback!\n");
-  exit(1);
-}
-void caml_js_wrap_callback_arguments () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback_arguments!\n");
-  exit(1);
-}
-void caml_js_wrap_callback_strict () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback_strict!\n");
-  exit(1);
-}
-void caml_js_wrap_meth_callback () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback!\n");
-  exit(1);
-}
-void caml_js_wrap_meth_callback_arguments () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_arguments!\n");
-  exit(1);
-}
-void caml_js_wrap_meth_callback_strict () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_strict!\n");
-  exit(1);
-}
-void caml_js_wrap_meth_callback_unsafe () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_unsafe!\n");
-  exit(1);
-}
 void caml_jsbytes_of_string () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_jsbytes_of_string!\n");
-  exit(1);
-}
-void caml_jsstring_of_string () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_jsstring_of_string!\n");
   exit(1);
 }
 void caml_list_mount_point () {
@@ -216,14 +124,6 @@ void caml_mount_autoload () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_mount_autoload!\n");
   exit(1);
 }
-void caml_ojs_new_arr () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_ojs_new_arr!\n");
-  exit(1);
-}
-void caml_pure_js_expr () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_pure_js_expr!\n");
-  exit(1);
-}
 void caml_read_file_content () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_read_file_content!\n");
   exit(1);
@@ -236,19 +136,11 @@ void caml_string_of_jsbytes () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_string_of_jsbytes!\n");
   exit(1);
 }
-void caml_string_of_jsstring () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_string_of_jsstring!\n");
-  exit(1);
-}
 void caml_unmount () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_unmount!\n");
   exit(1);
 }
 void caml_xmlhttprequest_create () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_xmlhttprequest_create!\n");
-  exit(1);
-}
-void debugger () {
-  fprintf(stderr, "Unimplemented Javascript primitive debugger!\n");
   exit(1);
 }
